requirements.txt
opencv-python
fastapi
Jinja2
uvicorn
pyyaml

# dev
black
flake8
isort